Flux/Lock interface changed, added error checking, changed comments, fixed #5

+++ src/Flux/Lock/Lock.php
@@ -3,14 +3,10 @@ declare(strict_types=1);
 
 namespace Flux\Lock;
 
-/**
- * Class Lock
- * @package Flux\Lock
- */
 class Lock implements LockInterface
 {
 
-    protected mixed $lockfilepointer = null;    // ?resource
+    protected mixed $lockfilepointer = false;    // false|resource, result from fopen()
     protected string $lockfilename = '';
     protected bool $locked = false;
 
@@ -18,9 +14,6 @@ class Lock implements LockInterface
     {
     }
 
-    /**
-     * @return bool
-     */
     public function acquire(): bool
     {
 
@@ -31,6 +24,10 @@ class Lock implements LockInterface
 
         $this->lockfilepointer = fopen($this->lockfilename, "w+");
 
+        // cannot open file
+        if ($this->lockfilepointer === false)
+            return false;
+
         $this->locked = flock($this->lockfilepointer, LOCK_EX | LOCK_NB);
 
         if (!$this->locked)
@@ -40,14 +37,17 @@ class Lock implements LockInterface
 
     }
 
-    public function release()
+    public function release(): void
     {
         if ($this->locked) {
             flock($this->lockfilepointer, LOCK_UN); // release the lock
             fclose($this->lockfilepointer);
         }
 
-        unlink($this->lockfilename);
+        if ($this->lockfilepointer !== false) {
+            unlink($this->lockfilename);
+        }
+
     }
